Fixed two problems: Oct-base sizes and reordered data.
The leading zeros in the size field were causing the emulator to think that
the requested size was in base 8 (as per syntax goes in C-like languages), so
now the leading zeros are removed when seen.
Also, when the resulting .mem files were merged, they were placed in order
according to their names, rather than their offsets in the emulated memory, so
now the files are prefixed with an increasing counter while looping over the
variables in the order they are positioned in memory.